chatbot-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
openai==1.57.0
google-generativeai==0.8.3
qdrant-client==1.12.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic[email]==2.10.3
langchain==0.3.9
langchain-openai==0.2.9
langchain-qdrant==0.2.0
tiktoken==0.8.0
numpy==1.26.4
python-multipart==0.0.20
httpx==0.28.1
passlib==1.7.4
bcrypt==5.0.0
python-jose[cryptography]==3.5.0

